backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
google-generativeai==0.8.5
resend==2.5.0
slowapi==0.1.9
python-dotenv==1.1.0
pydantic[email]==2.11.1
pydantic-settings==2.8.1
ruff==0.11.2
